requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-dateutil==2.8.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
WTForms==2.3.3
